Deplacer la logic depuis Game vers EntityActivity pour qu'on puisse faire un save/load et ajouter la fonction de save and load (appuie sur la lettre S fait un save, et appuier sur la letter L fait un load depuis la dérniere sauveguard)
