+++ Game_of_life_map/entity/bob.py
@@ -11,7 +11,6 @@ class Bob (Entity):
         self.set_initial_position()
         self.speed = 1
         self.target = self.get_target()
-        self.font = pygame.font.Font(None, 36)
         self.perception = 0
         self.mass = 1
