+++ Game_of_life_map/logic_game/entity_activity.py
@@ -6,9 +6,10 @@ from constant.settings import*
 
 
 class EntityActivity :
-    def __init__ (self, list_bob, dict_food):
-        self.list_bob = list_bob
-        self.dict_food = dict_food
+
+    def __init__ (self):
+        self.list_bob = self.create_list_bob()
+        self.dict_food = self.create_dict_food()
     
     @staticmethod
     def check_collision (entity_1, entity_2):
@@ -16,6 +17,29 @@ class EntityActivity :
             return True
         return False
 
+    def create_list_bob (self):
+        bob_list = []
+        for _ in range(NUMBER_BOB):
+            bob = Bob()
+            bob_list.append(bob) 
+        return bob_list  
+
+    def create_dict_food (self):
+        dict_food = {}
+        for _ in range (NUMBER_FOOD):
+            food = Food()
+            food.set_position()
+            position = (food.grid_x, food.grid_y)
+            if  position in dict_food:
+                dict_food[position].energy += 100
+            else:
+                dict_food[position] = food
+            return dict_food
+
+    def update_move_bob (self):
+        for bob in self.list_bob:
+            bob.move_towards_target()
+            
     def bob_eat_food(self):
         keys_to_remove = []
         for bob in self.list_bob:
