+++ Game_of_life_map/object/game.py
@@ -9,19 +9,21 @@ from .common import *
 from entity.food import Food
 from entity.entity import Entity
 from logic_game.entity_activity import EntityActivity
+import pickle
+import codecs
 
 class Game:
-    
+
+
     def __init__ (self, screen, clock):
+        self.BOB_FONT = pygame.font.Font(None, 36)
         self.screen = screen
         self.clock = clock
         self.width, self.height = self.screen.get_size()
         self.camera = Camera(self.width, self.height)
         self.map = Map(GRID_SIZE, GRID_SIZE)
         self.map.render_map()
-        self.list_bob = self.create_list_bob()
-        self.dict_food = self.create_dict_food()
-        self.entity_activity = EntityActivity(self.list_bob, self.dict_food)
+        self.entity_activity = EntityActivity()
         self.tick = 0
         self.day = 0
 
@@ -52,10 +54,26 @@ class Game:
                 pygame.quit()
                 sys.exit()
             if event.type == pygame.KEYDOWN:
+                if event.key == pygame.K_s:
+                    self.save()
+                if event.key == pygame.K_l:
+                    self.load()
                 if event.key == pygame.K_ESCAPE:
                     pygame.quit()
                     sys.exit()
 
+    def load(self):
+         file = open('save.txt', 'rb')
+         s = file.read()
+         self.entity_activity = pickle.loads(codecs.decode(s,"base64"))
+
+    def save(self):
+        s = codecs.encode(pickle.dumps(self.entity_activity), "base64").decode()
+        file = open('save.txt', 'w')
+        file.write(s)
+        file.close()
+
+    
     def draw (self):
         scroll = self.camera.scroll
         self.screen.fill((0,0,0))
@@ -69,42 +87,22 @@ class Game:
     def draw_bob(self):
         map_block_tiles = self.map.block_tiles
         scroll = self.camera.scroll
-        for bob in self.list_bob :
+        for bob in self.entity_activity.list_bob :
             render_pos = get_render_pos(bob.grid_x, bob.grid_y)
             self.screen.blit(bob.get_scaled_bob(), (render_pos[0] + map_block_tiles.get_width()/2 + scroll.x,
                                                      render_pos[1] + map_block_tiles.get_height()/4 + scroll.y))
-            energy_text = bob.font.render(f'E: {bob.energy}, P: {bob.perception}', True, (255, 255, 255))
+            energy_text = self.BOB_FONT.render(f'E: {bob.energy}, P: {bob.perception}', True, (255, 255, 255))
             text_rect = energy_text.get_rect(center=(render_pos[0] + map_block_tiles.get_width()/2 + scroll.x,
                                                   render_pos[1] + map_block_tiles.get_height()/4 + scroll.y - 20))
             self.screen.blit(energy_text, text_rect)
         
-    def create_list_bob (self):
-        bob_list = []
-        for _ in range(NUMBER_BOB):
-            bob = Bob()
-            bob_list.append(bob) 
-        return bob_list  
-
     def update_move_bob (self):
-        for bob in self.list_bob:
-            bob.move_towards_target()
-        
-    def create_dict_food (self):
-            dict_food = {}
-            for _ in range (NUMBER_FOOD):
-                food = Food()
-                food.set_position()
-                position = (food.grid_x, food.grid_y)
-                if  position in dict_food:
-                    dict_food[position].energy += 100
-                else:
-                    dict_food[position] = food
-            return dict_food
+        self.entity_activity.update_move_bob()
     
     def draw_food(self):
         map_block_tiles = self.map.block_tiles
         scroll = self.camera.scroll
-        for food in self.dict_food.values():
+        for food in self.entity_activity.dict_food.values():
             render_pos = get_render_pos (food.grid_x, food.grid_y)
             self.screen.blit (food.get_scaled_food(), (render_pos[0] + map_block_tiles.get_width()/2 + scroll.x, render_pos[1] + map_block_tiles.get_height()/4 + scroll.y))
     
@@ -123,10 +121,8 @@ class Game:
     # bob die : lost all of energy
     def die(self):
         self.entity_activity.die()
-
     
     def draw_text(self):
-
         font = pygame.font.SysFont(None, 50)
         text_surface = font.render('tick={}'.format(self.tick), True, (255,255,255))
         text_surface_day = font.render('day={}'.format(self.day), True, (255,255,255))
